On serving posts, recompile the project.
This custom HTTP server for developing is a copy of Python's HTTP server
with a twist: if a request to a post is received (trigger
'out/site/articles/**/*.html'), the server will recompile the site
(calling 'make') and only then it will respond the request.

If the recompilation fails, a HTTP error is sent.

With this, we can recompile the site on demand pressing F5 from the
browser without either recompiling the site by hand (calling make) nor
implementing a kind of watch-for-changes in Tup/Make like other
blog-engines do.

While nice, they are terrible for performance because
they trigger a recompile on *every change*, no matter what or when. With
the implemented server we know when to recompile: when the developer is
looking for a page change pressing F5.