Fix fig and impl notes (j2). Impl hook call from j2.
The hook call allows to wrap code in a code-fenced block, pass it
unchanged through Pandoc's Markdown parsing and interpret it in a
Panflute hook.

There, the code can be replaced by arbitrary code like Markdown to be
processed by Pandoc *after* Pandoc processed the whole file but *before*
outputing the HTML.

In this way we can have a "separated" Markdown context. Currently this
is used to interpret Markdown text in figures' captions and margin notes
and produce correct HTML which Pandoc cannot do if the Markdown is
interpreted with the rest of the file.