Don't glob over out/ folder because it may not exist all the files yet.
This is the limitation of our glob() function: it scans the real
filesystem so if a file does not exist yet, it will not be returned by
glob().

Using then glob('out/....') may not return a particular output file
because by the moment the Tupfile is created the file may not exist.

This contrasts with tup's glob implementation where it does not scan the
filesystem but returns any pattern that was used in the Tupfile (because
tup knows that eventually the files listed there will exist).

The solution is to keep using out glob() but ensuring that we always
scan a source folder (like z/ and posts/) and never an output folder
(like out/).